+++ src/main/logutils.ts
@@ -1,6 +1,6 @@
 /* eslint import/prefer-default-export: off, import/no-mutable-exports: off */
 import { Combatant } from './combatant';
-import { startRecording, stopRecording, isRecording }  from './main';
+import { recorder }  from './main';
 import { battlegrounds }  from './constants';
 
 const tail = require('tail').Tail;
@@ -12,34 +12,14 @@ let tailHandler: any;
 let currentLogFile: string;
 let lastLogFile: string;
 let videoStartDate: Date;
+let metadata: Metadata;
+let combatantMap: Map<string, Combatant> = new Map();
+let playerCombatant: Combatant | undefined;
 
 /**
- * Is wow running? Starts false but we'll check immediately on start-up. 
+ * wowProcessStopped
  */
-let isWowRunning: boolean = false;
-
-const wowProcessStarted = () => {
-    console.log("Wow.exe has started");
-    isWowRunning = true;
-};
-
-const wowProcessStopped = () => {
-    console.log("Wow.exe has stopped");
-    isWowRunning = false;
-    if (!isRecording) return; 
-
-    const videoStopDate = new Date();
-    const milliSeconds = (videoStopDate.getTime() - videoStartDate.getTime()); 
-    metadata.duration = Math.round(milliSeconds / 1000);
-
-    // Assume loss as game was closed. 
-    metadata.result = false;
-
-    stopRecording(metadata);
-};
-
-
-type Metadata = {
+ type Metadata = {
     name: string;
     category: string;
     zoneID?: number;
@@ -52,9 +32,39 @@ type Metadata = {
     teamMMR?: number;
 }
 
-let metadata: Metadata;
-let combatantMap: Map<string, Combatant> = new Map();
-let playerCombatant: Combatant | undefined;
+/**
+ * Is wow running? Starts false but we'll check immediately on start-up. 
+ */
+let isWowRunning: boolean = false;
+
+/**
+ * wowProcessStarted
+ */
+const wowProcessStarted = () => {
+    console.log("Wow.exe has started");
+    isWowRunning = true;
+    recorder.startBuffer();
+};
+
+/**
+ * wowProcessStopped
+ */
+const wowProcessStopped = () => {
+    console.log("Wow.exe has stopped");
+    isWowRunning = false;
+
+    if (recorder.isRecording) {
+        const videoStopDate = new Date();
+        const milliSeconds = (videoStopDate.getTime() - videoStartDate.getTime()); 
+        metadata.duration = Math.round(milliSeconds / 1000);
+    
+        // Assume loss as game was closed. 
+        metadata.result = false;
+        recorder.stop(metadata);
+    } else if (recorder.isRecordingBuffer) {
+        recorder.stopBuffer();
+    }
+};
 
 /**
  * getLatestLog 
@@ -125,7 +135,10 @@ const handleLogLine = (line: string) => {
 const handleArenaStartLine = (line: string) => {
     const zoneID = parseInt(line.split(',')[1]);
     const category = line.split(',')[3];
-    videoStartDate = new Date();
+
+    // If all goes to plan we don't need this but we do it incase the game
+    // crashes etc. so we can still get a reasonable duration.
+    videoStartDate = getCombatLogDate(line);
 
     metadata = {
         name: "name",
@@ -134,15 +147,15 @@ const handleArenaStartLine = (line: string) => {
         duration: 0,
         result: false,
     }
-
-    startRecording(metadata);
+    
+    recorder.start();
 }
 
 /**
  * Handle a line from the WoW log. 
  */
  const handleArenaStopLine = (line: string) => {
-    if (!isRecording) return; 
+    if (!recorder.isRecording) return; 
 
     if (playerCombatant) {
         metadata.playerName = playerCombatant.name;
@@ -150,10 +163,12 @@ const handleArenaStartLine = (line: string) => {
         metadata.playerSpecID = playerCombatant.specID;        
     }
 
-    const videoStopDate = new Date();
-    const milliSeconds = (videoStopDate.getTime() - videoStartDate.getTime()); 
-    const duration = Math.round(milliSeconds / 1000);
-    const [result, MMR] = determineArenaMatchResult(line);
+    // Add a few seconds so we reliably can see the end screen.
+    const overrun = 3;
+
+    // Helpfully ARENA_MATCH_END events contain the game duration. 
+    const duration = parseInt(line.split(',')[2]) + overrun;
+    const [result, MMR] = determineArenaMatchResult(line);   
 
     metadata.duration = duration; 
     metadata.result = result;
@@ -161,7 +176,8 @@ const handleArenaStartLine = (line: string) => {
 
     combatantMap.clear();
     playerCombatant = undefined;
-    stopRecording(metadata);
+
+    recorder.stop(metadata, overrun);
 }
 
 /**
@@ -194,7 +210,8 @@ const determineArenaMatchResult = (line: string): any[] => {
  const handleRaidStartLine = (line: string) => {
     const encounterID = parseInt(line.split(',')[1]);
     const category = "Raids";
-    videoStartDate = new Date();
+
+    videoStartDate = getCombatLogDate(line);
 
     metadata = {
         name: "name",
@@ -204,14 +221,14 @@ const determineArenaMatchResult = (line: string): any[] => {
         result: false,
     }
 
-    startRecording(metadata);
+    recorder.start();
 }
 
 /**
  * Handle a line from the WoW log. 
  */
  const handleRaidStopLine = (line: string) => {
-    if (!isRecording) return; 
+    if (!recorder.isRecording) return; 
 
     if (playerCombatant) {
         metadata.playerName = playerCombatant.name;
@@ -219,16 +236,20 @@ const determineArenaMatchResult = (line: string): any[] => {
         metadata.playerSpecID = playerCombatant.specID;        
     }
 
-    const videoStopDate = new Date();
+    // Add a few seconds so we reliably see the aftermath of a kill.
+    const overrun = 15;
+
+    const videoStopDate = getCombatLogDate(line);
     const milliSeconds = (videoStopDate.getTime() - videoStartDate.getTime()); 
-    const duration = Math.round(milliSeconds / 1000);
+    const duration = Math.round(milliSeconds / 1000) + overrun;
 
     metadata.duration = duration; 
     metadata.result = determineRaidEncounterResult(line);
     
     combatantMap.clear();
     playerCombatant = undefined;
-    stopRecording(metadata);
+
+    recorder.stop(metadata, overrun);
 }
 
 /**
@@ -239,15 +260,15 @@ const determineArenaMatchResult = (line: string): any[] => {
     const zoneID = parseInt(line.split(',')[1]);
     const isBG = battlegrounds.hasOwnProperty(zoneID);
 
-    if (!isRecording && isBG) {
+    if (!recorder.isRecording && isBG) {
         console.log("ZONE_CHANGE into BG, start recording");
-        battlegroundStartRecording(line);   
-    } else if (isRecording && !isBG ) {
+        battlegroundStart(line);   
+    } else if (recorder.isRecording && !isBG ) {
         console.log("ZONE_CHANGE out of BG, stop recording");
-        battlegroundStopRecording();
-    } else if (isRecording && !isBG) {
+        battlegroundStop(line);
+    } else if (recorder.isRecording && !isBG) {
         console.log("ZONE_CHANGE out of unknown content, stop recording");
-        zoneChangeStopRecording();
+        zoneChangeStop(line);
     }
 }
 
@@ -290,11 +311,12 @@ const handleCombatantInfoLine = (line: string) => {
 /**
  * ZONE_CHANGE event into a BG.  
  */
- const battlegroundStartRecording = (line: string) => {
+ const battlegroundStart = (line: string) => {
     const zoneID = parseInt(line.split(',')[1]);
     const battlegroundName = battlegrounds[zoneID];
     const category = "Battlegrounds";
-    videoStartDate = new Date();
+
+    videoStartDate = getCombatLogDate(line);
 
     metadata = {
         name: battlegroundName,
@@ -304,34 +326,34 @@ const handleCombatantInfoLine = (line: string) => {
         result: false,
     }
 
-    startRecording(metadata);
+    recorder.start();
 }
 
 /**
- * battlegroundStopRecording
+ * battlegroundStop
  */
- const battlegroundStopRecording = () => {
-    const videoStopDate = new Date();
+ const battlegroundStop = (line: string) => {
+    const videoStopDate = getCombatLogDate(line);
     const milliSeconds = (videoStopDate.getTime() - videoStartDate.getTime()); 
     metadata.duration = Math.round(milliSeconds / 1000);
 
     // No idea how we can tell who has won a BG so assume loss. 
     // I've just disabled displaying this in the UI so this does nothing.
     metadata.result = false;
-    stopRecording(metadata);
+    recorder.stop(metadata);
 }
 
 /**
- * zoneChangeStopRecording
+ * zoneChangeStop
  */
- const zoneChangeStopRecording = () => {
-    const videoStopDate = new Date();
+ const zoneChangeStop = (line: string) => {
+    const videoStopDate = getCombatLogDate(line);
     const milliSeconds = (videoStopDate.getTime() - videoStartDate.getTime()); 
     metadata.duration = Math.round(milliSeconds / 1000);
 
     // Assume loss if zoned out of content. 
     metadata.result = false;
-    stopRecording(metadata);
+    recorder.stop(metadata);
 }
 
 /**
@@ -417,6 +439,24 @@ const pollWowProcess = () => {
             wowProcessStopped();
         }
     }, 5000);
+}
+
+/**
+ * getCombatLogDate
+ */
+const getCombatLogDate = (line: string) => {
+    const [date, time] = line.split(" ");
+    const [month, day] = date.split("/");
+    const [hours, mins, secs] = time.split(":");
+
+    const dateObj = new Date();
+    dateObj.setDate(parseInt(day));
+    dateObj.setMonth(parseInt(month));
+    dateObj.setHours(parseInt(hours));
+    dateObj.setMinutes(parseInt(mins));
+    dateObj.setSeconds(parseInt(secs));
+
+    return dateObj;
 }
 
 export {