+++ src/renderer/Status.tsx
@@ -1,15 +1,18 @@
 import recordIcon from '../../assets/icon/record-icon.png';
 import eyeIcon from  '../../assets/icon/sleep-icon.png';
 import errorIcon from  '../../assets/icon/error-icon.png';
+import watchIcon from  '../../assets/icon/watch-icon.png';
 import * as React from 'react';
 
 export default function Status() {
 
   /**
-   * Python subprocess status.
+   * Recorder subprocess status.
    *   0 - waiting
    *   1 - recording
    *   2 - stopped/error
+   *   3 - watching
+   *   4 - saving
    */
   const [status, setStatus] = React.useState(0);
 
@@ -26,19 +29,27 @@ export default function Status() {
   function getStatus() {
     if (status === 0) {
       return(
-        <div id="status" title="Waiting for a game to start.">
-          <img id="eye-icon" alt="icon" src={eyeIcon}/>
+        <div id="status" title="Waiting for WoW to start">
+          <img id="eye-icon" alt="icon" src={ eyeIcon }/>
         </div>
       )} else if (status === 1) {
         return(
           <div id="status">
-            <img id="status-icon" title="Recording a game." alt="icon" src={recordIcon}/>
+            <img id="status-icon" title="Recording" alt="icon" src={ recordIcon }/>
           </div>
       )} else if (status === 2) {
         return(
           <div id="status">
-            <img id="error-icon" title="Failed to launch, fix config and restart." alt="icon" src={errorIcon}/>
+            <img id="error-icon" title="Failed to launch, fix config and restart" alt="icon" src={ errorIcon }/>
           </div>
+      )} else if (status === 3) {
+        return(
+          <div id="status">
+            <img id="error-icon" title="Ready and waiting" alt="icon" src={ watchIcon }/>
+          </div>
+      )} else {
+        return(
+          <div>Never reach here, but avoid linter error.</div>
       )}
     }